chore: port to typescript
