+++ .eslintrc.json
@@ -1,10 +1,18 @@
 {
+    "root": true,
+    "parser": "@typescript-eslint/parser",
+    "plugins": [
+        "@typescript-eslint"
+    ],
     "env": {
         "commonjs": true,
         "es6": true,
         "node": true
     },
     "extends": [
+        "eslint:recommended",
+        "plugin:@typescript-eslint/eslint-recommended",
+        "plugin:@typescript-eslint/recommended",
         "google"
     ],
     "globals": {
