+++ index.ts
@@ -1,7 +1,8 @@
-const docker = require('./docker');
-const humanitecFactory = require('./humanitec');
-const fs = require('fs');
-const core = require('@actions/core');
+import * as docker from './docker';
+import {humanitecFactory} from './humanitec';
+
+import {existsSync} from 'node:fs';
+import * as core from '@actions/core';
 
 /**
  * Performs the GitHub action.
@@ -10,18 +11,18 @@ async function runAction() {
   // Get GitHub Action inputs
   const token = core.getInput('humanitec-token', {required: true});
   const orgId = core.getInput('organization', {required: true});
-  const imageName = core.getInput('image-name') || process.env.GITHUB_REPOSITORY.replace(/.*\//, '');
+  const imageName = core.getInput('image-name') || (process.env.GITHUB_REPOSITORY || '').replace(/.*\//, '');
   const context = core.getInput('context') || core.getInput('dockerfile') || '.';
   const file = core.getInput('file') || '';
   const registryHost = core.getInput('humanitec-registry') || 'registry.humanitec.io';
   const apiHost = core.getInput('humanitec-api') || 'api.humanitec.io';
   const tag = core.getInput('tag') || '';
-  const commit = process.env.GITHUB_SHA;
+  const commit = process.env.GITHUB_SHA || '';
   const autoTag = /^\s*(true|1)\s*$/i.test(core.getInput('auto-tag'));
   const additionalDockerArguments = core.getInput('additional-docker-arguments') || '';
 
-  const ref = process.env.GITHUB_REF;
-  if (!fs.existsSync(`${process.env.GITHUB_WORKSPACE}/.git`)) {
+  const ref = process.env.GITHUB_REF || '';
+  if (!existsSync(`${process.env.GITHUB_WORKSPACE}/.git`)) {
     core.error('It does not look like anything was checked out.');
     core.error('Did you run a checkout step before this step? ' +
       'http:/docs.humanitec.com/connecting-your-ci#github-actions');
@@ -29,13 +30,13 @@ async function runAction() {
     return;
   }
 
-  if (file != '' && !fs.existsSync(file)) {
+  if (file != '' && !existsSync(file)) {
     core.error(`Cannot find file ${file}`);
     core.setFailed('Cannot find file.');
     return;
   }
 
-  if (!fs.existsSync(context)) {
+  if (!existsSync(context)) {
     core.error(`Context path does not exist: ${context}`);
     core.setFailed('Context path does not exist.');
     return;
@@ -59,10 +60,10 @@ async function runAction() {
     return;
   }
 
-  process.chdir(process.env.GITHUB_WORKSPACE);
+  process.chdir((process.env.GITHUB_WORKSPACE || ''));
 
   let version = '';
-  if (autoTag && ref.includes('\/tags\/')) {
+  if (autoTag && ref.includes('/tags/')) {
     version = ref.replace(/.*\/tags\//, '');
   } else if (tag) {
     version = tag;
