+++ package.json
@@ -2,18 +2,24 @@
   "name": "build-push-to-humanitec",
   "version": "1.0.0",
   "description": "GitHub Action for easy integration with Humanitec",
-  "main": "index.js",
+  "main": "dist/index.js",
   "scripts": {
-    "lint": "eslint index.js docker.js humanitec.js",
-    "package": "ncc build index.js -o dist"
+    "lint": "eslint index.ts docker.ts humanitec.ts",
+    "package": "ncc build index.ts -o dist"
   },
   "author": "chris.stephenson@humanitec.com",
   "license": "ISC",
   "devDependencies": {
+    "@tsconfig/node16": "^1.0.3",
+    "@types/node": "^16.18.3",
+    "@types/node-fetch": "^2.6.2",
+    "@typescript-eslint/eslint-plugin": "^5.44.0",
+    "@typescript-eslint/parser": "^5.44.0",
     "@vercel/ncc": "^0.34.0",
-    "eslint": "^8.16.0",
+    "eslint": "^8.28.0",
     "eslint-config-eslint": "^7.0.0",
-    "eslint-config-google": "^0.14.0"
+    "eslint-config-google": "^0.14.0",
+    "typescript": "^4.9.3"
   },
   "dependencies": {
     "@actions/core": "^1.9.1",
