+++ tsconfig.json
@@ -1,0 +1,4 @@
+{
+  "extends": "@tsconfig/node16/tsconfig.json",
+  "include": ["*.ts"]
+}
